apps/gitloader-flask/requirements.txt
numpy==1.24.3
Flask==2.2.2
langchain==0.0.149
python-dotenv==1.0.0
supabase==1.0.3
tiktoken==0.3.3
